Week1/Assignment1: How does a browser works?
Exercise1.1:1.When a user enters an URL in the browser, how does the browser fetch the desired result ? Explain this with the below in mind and Demonstrate this by drawing a diagram for the same.(2-3hours)

a. What is the main functionality of the browser?
b. High Level Components of a browser.
c. Rendering engine and its use.
d. Parsers (HTML, CSS, etc)
e. Script Processors
f. Tree construction
g. Order of script processing
h. Layout and Painting